changed Long to Int for Database ID

+++ Source/src/za/co/zebrav/smartdoor/AddVoiceFragment.java
@@ -34,7 +34,7 @@ public class AddVoiceFragment extends ListFragment implements OnClickListener
 	public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
 	{
 		Bundle bundle = this.getArguments();
-		Long id = bundle.getLong("userID", -1);
+		Integer id = bundle.getInt("userID", -1);
 		activeKey = id.toString();
 		return inflater.inflate(R.layout.fragment_add_voice, container, false);
 	}
